backend/requirements.txt
Django==4.0.4
djangorestframework==3.13.1
drf-extensions==0.7.1
pytest
pytest-django
model_bakery